requirements.txt
django==2.2.13
django-auditlog==0.4.7
django-braces==1.13.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.2.5
djangorestframework==3.11.0
django-filter==2.2.0
django-filters==0.2.1
django-handyhelpers==0.0.22
django-userextensions==0.0.7
drf-dynamic-fields==0.3.1
djangorestframework-filters==1.0.0.dev0
django-rest-swagger==2.2.0
drf-dynamic-fields==0.3.1
requests==2.24.0
